Format code with pep8

+++ item_catalog/vagrant/catalog/project.py
@@ -1,20 +1,18 @@
-from flask import Flask, render_template, request, redirect,jsonify, url_for, flash
+from flask import Flask, render_template, request
+from flask import redirect, jsonify, url_for, flash
 from functools import wraps
-app = Flask(__name__)
-
 from sqlalchemy import create_engine, asc
 from sqlalchemy.orm import sessionmaker
 from database_setup import Base, Category, Item, User
-
 from flask import session as login_session
-import random, string
-
 from oauth2client.client import flow_from_clientsecrets
 from oauth2client.client import FlowExchangeError
 import httplib2
 import json
 from flask import make_response
 import requests
+import random
+import string
 
 app = Flask(__name__)
 
@@ -22,7 +20,7 @@ CLIENT_ID = json.loads(
     open('client_secrets.json', 'r').read())['web']['client_id']
 APPLICATION_NAME = "Catalog Item Application"
 
-#Connect to Database and create database session
+# Connect to Database and create database session
 engine = create_engine('sqlite:///categoryitemwithusers.db')
 Base.metadata.bind = engine
 
@@ -51,8 +49,9 @@ def fbconnect():
         'web']['app_id']
     app_secret = json.loads(
         open('fb_client_secrets.json', 'r').read())['web']['app_secret']
-    url = 'https://graph.facebook.com/oauth/access_token?grant_type=fb_exchange_token&client_id=%s&client_secret=%s&fb_exchange_token=%s' % (
-        app_id, app_secret, access_token)
+    url = ('https://graph.facebook.com/oauth/access_token?grant_type=' +
+           'fb_exchange_token&client_id=%s&client_secret=%s&' +
+           'fb_exchange_token=%s') % (app_id, app_secret, access_token)
     h = httplib2.Http()
     result = h.request(url, 'GET')[1]
 
@@ -60,8 +59,6 @@ def fbconnect():
     userinfo_url = "https://graph.facebook.com/v2.8/me"
     # strip expire tag from access token
     token = result.split("&")[0]
-
-
     url = 'https://graph.facebook.com/v2.8/me?%s&fields=name,id,email' % token
     h = httplib2.Http()
     result = h.request(url, 'GET')[1]
@@ -73,12 +70,12 @@ def fbconnect():
     login_session['email'] = data["email"]
     login_session['facebook_id'] = data["id"]
 
-    # The token must be stored in the login_session in order to properly logout, let's strip out the information before the equals sign in our token
     stored_token = token.split("=")[1]
     login_session['access_token'] = stored_token
 
     # Get user picture
-    url = 'https://graph.facebook.com/v2.8/me/picture?%s&redirect=0&height=200&width=200' % token
+    url = ('https://graph.facebook.com/v2.8/me/picture?' +
+           '%s&redirect=0&height=200&width=200' % token)
     h = httplib2.Http()
     result = h.request(url, 'GET')[1]
     data = json.loads(result)
@@ -98,7 +95,8 @@ def fbconnect():
     output += '!</h1>'
     output += '<img src="'
     output += login_session['picture']
-    output += ' " style = "width: 300px; height: 300px;border-radius: 150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
+    output += ' " style = "width: 300px; height: 300px;border-radius: 150px;'
+    output += '-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
 
     flash("Now logged in as %s" % login_session['username'])
     return output
@@ -157,8 +155,8 @@ def gconnect():
     stored_access_token = login_session.get('access_token')
     stored_gplus_id = login_session.get('gplus_id')
     if stored_access_token is not None and gplus_id == stored_gplus_id:
-        response = make_response(json.dumps('Current user is already connected.'),
-                                 200)
+        response = make_response(json.dumps(
+                                'Current user is already connected.'), 200)
         response.headers['Content-Type'] = 'application/json'
         return response
 
@@ -191,7 +189,8 @@ def gconnect():
     output += '!</h1>'
     output += '<img src="'
     output += login_session['picture']
-    output += ' " style = "width: 300px; height: 300px;border-radius: 150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
+    output += ' " style = "width: 300px; height: 300px;border-radius: 150px;'
+    output += '-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
     flash("you are now logged in as %s" % login_session['username'])
     print "done!"
     return output
@@ -204,16 +203,19 @@ def gdisconnect():
     access_token = login_session['access_token']
     if access_token is None:
         print 'Access Token is None'
-        response = make_response(json.dumps('Current user not connected.'), 401)
+        response = make_response(json.dumps
+                                 ('Current user not connected.'), 401)
         response.headers['Content-Type'] = 'application/json'
         return response
-    url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' % login_session['access_token']
+    url = ('https://accounts.google.com/o/oauth2/' +
+           'revoke?token=%s') % login_session['access_token']
     h = httplib2.Http()
     result = h.request(url, 'GET')[0]
     print 'result is '
     print result
     if result['status'] != '200':
-        response = make_response(json.dumps('Failed to revoke token for given user.', 400))
+        response = make_response(json.dumps("""Failed to revoke token
+                                 for given user.""", 400))
         response.headers['Content-Type'] = 'application/json'
         return response
 
@@ -223,7 +225,8 @@ def fbdisconnect():
     facebook_id = login_session['facebook_id']
     # The access token must me included to successfully logout
     access_token = login_session['access_token']
-    url = 'https://graph.facebook.com/%s/permissions?access_token=%s' % (facebook_id,access_token)
+    url = ('https://graph.facebook.com/%s/permissions' +
+           '?access_token=%s') % (facebook_id, access_token)
     h = httplib2.Http()
     result = h.request(url, 'DELETE')[1]
     return "you have been logged out"
@@ -252,22 +255,25 @@ def disconnect():
         return redirect(url_for('showCategories'))
 
 
-#JSON APIs to view category Information
+# JSON APIs to view category Information
 @app.route('/category/JSON')
 def categoryJSON():
     categories = session.query(Category).all()
     return jsonify(categories=[c.serialize for c in categories])
 
+
 @app.route('/category/<int:category_id>/item/JSON')
 def categoryItemJSON(category_id):
-    category = session.query(Category).filter_by(id = category_id).one()
-    items = session.query(Item).filter_by(category_id = category_id).all()
+    category = session.query(Category).filter_by(id=category_id).one()
+    items = session.query(Item).filter_by(category_id=category_id).all()
     return jsonify(Items=[i.serialize for i in items])
+
 
 @app.route('/category/<int:category_id>/item/<int:item_id>/JSON')
 def itemJSON(category_id, item_id):
     i = session.query(Item).filter_by(id=item_id).one()
     return jsonify(Item=i.serialize)
+
 
 # Refuctor cheking login status
 def login_required(f):
@@ -280,6 +286,7 @@ def login_required(f):
             return redirect('/login')
     return decorated_function
 
+
 # Show all categories
 @app.route('/')
 @app.route('/category/')
@@ -290,13 +297,14 @@ def showCategories():
     else:
         return render_template('category.html', categories=categories)
 
-#Create a new category
-@app.route('/category/new/', methods=['GET','POST'])
+
+# Create a new category
+@app.route('/category/new/', methods=['GET', 'POST'])
 @login_required
 def newCategory():
     if request.method == 'POST':
         newcategory = Category(
-            name = request.form['name'], user_id=login_session['user_id'])
+            name=request.form['name'], user_id=login_session['user_id'])
         session.add(newcategory)
         flash('New Category %s Successfully Created' % newcategory.name)
         session.commit()
@@ -304,11 +312,15 @@ def newCategory():
     else:
         return render_template('newCategory.html')
 
-#Edit a category
-@app.route('/category/<int:category_id>/edit/', methods = ['GET', 'POST'])
+
+# Edit a category
+@app.route('/category/<int:category_id>/edit/', methods=['GET', 'POST'])
 @login_required
 def editCategory(category_id):
     editedcategory = session.query(Category).filter_by(id=category_id).one()
+    if login_session['user_id'] != editedcategory.user_id:
+        flash("You are not authorized to edit this category.")
+        return redirect('/category')
     if request.method == 'POST':
         if request.form['name']:
             editedcategory.name = request.form['name']
@@ -317,39 +329,51 @@ def editCategory(category_id):
     else:
         return render_template('editCategory.html', category=editedcategory)
 
-#Delete a category
-@app.route('/category/<int:category_id>/delete/', methods = ['GET','POST'])
+
+# Delete a category
+@app.route('/category/<int:category_id>/delete/', methods=['GET', 'POST'])
 @login_required
 def deleteCategory(category_id):
     categoryToDelete = session.query(Category).filter_by(id=category_id).one()
+    if login_session['user_id'] != categoryToDelete.user_id:
+        flash("You are not authorized to delete this category.")
+        return redirect('/category')
     if request.method == 'POST':
         session.delete(categoryToDelete)
         flash('%s Successfully Deleted' % categoryToDelete.name)
         session.commit()
         return redirect(url_for('showCategories'))
     else:
-        return render_template('deleteCategory.html',category=categoryToDelete, category_id=category_id)
+        return render_template('deleteCategory.html',
+                               category=categoryToDelete,
+                               category_id=category_id)
 
 
-#Show a category item
+# Show a category item
 @app.route('/category/<int:category_id>/')
 @app.route('/category/<int:category_id>/item/')
 def showItem(category_id):
     category = session.query(Category).filter_by(id=category_id).one()
     items = session.query(Item).filter_by(category_id=category_id).all()
     creator = getUserInfo(category.user_id)
-    if 'username' not in login_session or creator.id != login_session['user_id']:
-        return render_template('publicItem.html', items=items, category=category, creator=creator)
+    if ('username' not in login_session or
+       creator.id != login_session['user_id']):
+        return render_template('publicItem.html', items=items,
+                               category=category, creator=creator)
     else:
-        return render_template('item.html', items=items, category=category, creator=creator)
+        return render_template('item.html', items=items,
+                               category=category, creator=creator)
 
-#Create a new item
-@app.route('/category/<int:category_id>/item/new/',methods=['GET','POST'])
+
+# Create a new item
+@app.route('/category/<int:category_id>/item/new/', methods=['GET', 'POST'])
 @login_required
 def newItem(category_id):
-    category = session.query(Category).filter_by(id = category_id).one()
+    category = session.query(Category).filter_by(id=category_id).one()
     if request.method == 'POST':
-        new_item = Item(name=request.form['name'], description=request.form['description'], category_id=category_id, user_id=category.user_id)
+        new_item = Item(name=request.form['name'],
+                        description=request.form['description'],
+                        category_id=category_id, user_id=category.user_id)
         session.add(new_item)
         session.commit()
         flash('%s Successfully Created' % (new_item.name))
@@ -357,12 +381,17 @@ def newItem(category_id):
     else:
         return render_template('newItem.html', category_id=category_id)
 
-#Edit a item
-@app.route('/category/<int:category_id>/item/<int:item_id>/edit', methods=['GET','POST'])
+
+# Edit a item
+@app.route('/category/<int:category_id>/item/<int:item_id>/edit',
+           methods=['GET', 'POST'])
 @login_required
 def editItem(category_id, item_id):
     editedItem = session.query(Item).filter_by(id=item_id).one()
     category = session.query(Category).filter_by(id=category_id).one()
+    if login_session['user_id'] != category.user_id:
+        flash("You are not authorized to edit this item.")
+        return redirect(url_for('showItem', category_id=category_id))
     if request.method == 'POST':
         if request.form['name']:
             editedItem.name = request.form['name']
@@ -373,22 +402,28 @@ def editItem(category_id, item_id):
         flash('Item Successfully Edited')
         return redirect(url_for('showItem', category_id=category_id))
     else:
-        return render_template('editItem.html', category_id=category_id, item_id=item_id, item=editedItem)
+        return render_template('editItem.html', category_id=category_id,
+                               item_id=item_id, item=editedItem)
 
 
-#Delete a item
-@app.route('/category/<int:category_id>/item/<int:item_id>/delete', methods = ['GET','POST'])
+# Delete a item
+@app.route('/category/<int:category_id>/item/<int:item_id>/delete',
+           methods=['GET', 'POST'])
 @login_required
 def deleteItem(category_id, item_id):
     category = session.query(Category).filter_by(id=category_id).one()
     itemToDelete = session.query(Item).filter_by(id=item_id).one()
+    if login_session['user_id'] != category.user_id:
+        flash("You are not authorized to delete this item.")
+        return redirect(url_for('showItem', category_id=category_id))
     if request.method == 'POST':
         session.delete(itemToDelete)
         session.commit()
         flash('Item Successfully Deleted')
         return redirect(url_for('showItem', category_id=category_id))
     else:
-        return render_template('deleteItem.html', category_id=category_id, item_id=item_id, item=itemToDelete)
+        return render_template('deleteItem.html', category_id=category_id,
+                               item_id=item_id, item=itemToDelete)
 
 
 # User Helper Functions
@@ -399,6 +434,7 @@ def createUser(login_session):
     session.commit()
     user = session.query(User).filter_by(email=login_session['email']).one()
     return user.id
+
 
 def getUserInfo(user_id):
     user = session.query(User).filter_by(id=user_id).one()
@@ -414,6 +450,6 @@ def getUserID(email):
 
 
 if __name__ == '__main__':
-  app.secret_key = '6pKQC9BWFdQNtNHDkE'
-  app.debug = True
-  app.run(host = '0.0.0.0', port = 8000)
+    app.secret_key = '6pKQC9BWFdQNtNHDkE'
+    app.debug = True
+    app.run(host='0.0.0.0', port=8000)
